refactor(telegram): extract reusable utils and split history.ts
Move cosineDistance to src/utils/math.ts and parseDate to src/utils/date.ts
for cross-tool reuse. Split history.ts logic into lib/download.ts,
lib/search.ts, and lib/export.ts. Make handler store required instead
of optional — outgoing reply persistence is standard behavior.
