+++ src/telegram/commands/history.ts
@@ -1,30 +1,20 @@
 import { Command } from "commander";
 import * as p from "@clack/prompts";
 import pc from "picocolors";
-import logger from "@app/logger";
-import { TelegramToolConfig } from "../lib/TelegramToolConfig";
-import { TGClient } from "../lib/TGClient";
-import { TelegramMessage } from "../lib/TelegramMessage";
-import type { SerializedMessage } from "../lib/TelegramMessage";
-import { TelegramHistoryStore } from "../lib/TelegramHistoryStore";
-import type { ContactConfig, SearchResult } from "../lib/types";
-import { EMBEDDING_LANGUAGES } from "../lib/types";
+import { parseDate } from "@app/utils/date";
 import { formatNumber } from "@app/utils/format";
 import { formatTable } from "@app/utils/table";
 import { embedText, detectLanguage } from "@app/utils/macos/nlp";
-import type { EmbedResult } from "@app/utils/macos/types";
+import { TelegramToolConfig } from "../lib/TelegramToolConfig";
+import { TGClient } from "../lib/TGClient";
+import { TelegramHistoryStore } from "../lib/TelegramHistoryStore";
+import type { ContactConfig, SearchResult } from "../lib/types";
+import { EMBEDDING_LANGUAGES } from "../lib/types";
+import { downloadContact, embedMessages } from "../lib/download";
+import { displayResults } from "../lib/search";
+import { formatMessages, VALID_EXPORT_FORMATS, type ExportFormat } from "../lib/export";
 
 // ── Helpers ───────────────────────────────────────────────────────────
-
-function parseDate(value: string): Date {
-	const d = new Date(value);
-
-	if (Number.isNaN(d.getTime())) {
-		throw new Error(`Invalid date: ${value}`);
-	}
-
-	return d;
-}
 
 async function ensureClient(config: TelegramToolConfig): Promise<TGClient | null> {
 	if (!config.hasValidSession()) {
@@ -176,7 +166,6 @@ function registerDownloadCommand(history: Command): void {
 						const { embedded, skipped, unsupportedLangs } = await embedMessages(
 							store,
 							contact.userId,
-							contact.displayName,
 						);
 
 						embedSpinner.stop(
@@ -202,179 +191,7 @@ function registerDownloadCommand(history: Command): void {
 		});
 }
 
-async function downloadContact(
-	client: TGClient,
-	store: TelegramHistoryStore,
-	contact: ContactConfig,
-	options: { since?: Date; until?: Date; limit?: number },
-): Promise<void> {
-	p.log.step(pc.bold(contact.displayName));
-
-	const spinner = p.spinner();
-	spinner.start("Counting messages...");
-
-	let totalEstimate: number;
-
-	try {
-		totalEstimate = await client.getMessageCount(contact.userId);
-	} catch {
-		spinner.stop("Could not count messages");
-		totalEstimate = 0;
-	}
-
-	const lastSyncedId = store.getLastSyncedId(contact.userId);
-	const isIncremental = lastSyncedId !== null && !options.since;
-
-	if (isIncremental) {
-		spinner.stop(`Found ${formatNumber(totalEstimate)} total messages (incremental sync from #${lastSyncedId})`);
-	} else {
-		spinner.stop(`Found ${formatNumber(totalEstimate)} total messages`);
-	}
-
-	const iterOptions: {
-		limit?: number;
-		offsetDate?: number;
-		minId?: number;
-	} = {};
-
-	if (options.limit) {
-		iterOptions.limit = options.limit;
-	}
-
-	if (options.until) {
-		iterOptions.offsetDate = Math.floor(options.until.getTime() / 1000);
-	}
-
-	if (isIncremental && lastSyncedId !== null) {
-		iterOptions.minId = lastSyncedId;
-	}
-
-	const progressSpinner = p.spinner();
-	progressSpinner.start("Downloading messages...");
-
-	const batch: SerializedMessage[] = [];
-	let downloaded = 0;
-	let inserted = 0;
-	let highestId = lastSyncedId ?? 0;
-	let retryCount = 0;
-	const BATCH_SIZE = 100;
-	const MAX_RETRIES = 5;
-
-	try {
-		for await (const apiMessage of client.getMessages(contact.userId, iterOptions)) {
-			const msg = new TelegramMessage(apiMessage);
-
-			if (options.since && msg.date < options.since) {
-				continue;
-			}
-
-			if (options.until && msg.date > options.until) {
-				continue;
-			}
-
-			batch.push(msg.toJSON());
-			downloaded++;
-
-			if (msg.id > highestId) {
-				highestId = msg.id;
-			}
-
-			if (batch.length >= BATCH_SIZE) {
-				const batchInserted = store.insertMessages(contact.userId, batch);
-				inserted += batchInserted;
-				batch.length = 0;
-				retryCount = 0;
-
-				progressSpinner.message(
-					`Downloaded ${formatNumber(downloaded)} messages (${formatNumber(inserted)} new)`,
-				);
-			}
-		}
-	} catch (err) {
-		const errorStr = String(err);
-
-		if (errorStr.includes("FLOOD_WAIT") || errorStr.includes("FloodWait")) {
-			const waitMatch = errorStr.match(/(\d+)/);
-			const waitSeconds = waitMatch ? parseInt(waitMatch[1], 10) : 30;
-			retryCount++;
-
-			if (retryCount <= MAX_RETRIES) {
-				const backoff = waitSeconds * Math.pow(2, retryCount - 1);
-				progressSpinner.message(
-					`Rate limited — waiting ${backoff}s (retry ${retryCount}/${MAX_RETRIES})`,
-				);
-				await Bun.sleep(backoff * 1000);
-			} else {
-				progressSpinner.stop(`Rate limited after ${MAX_RETRIES} retries`);
-				p.log.warn("Stopped due to persistent rate limiting. Run again later to resume.");
-			}
-		} else {
-			progressSpinner.stop("Error during download");
-			p.log.error(`Download error: ${errorStr}`);
-		}
-	}
-
-	if (batch.length > 0) {
-		const batchInserted = store.insertMessages(contact.userId, batch);
-		inserted += batchInserted;
-	}
-
-	if (highestId > 0) {
-		store.setLastSyncedId(contact.userId, highestId);
-	}
-
-	progressSpinner.stop(
-		`${pc.green(formatNumber(downloaded))} downloaded, ${pc.green(formatNumber(inserted))} new messages stored`,
-	);
-}
-
-// ── Embed ─────────────────────────────────────────────────────────────
-
-async function embedMessages(
-	store: TelegramHistoryStore,
-	chatId: string,
-	_displayName: string,
-): Promise<{ embedded: number; skipped: number; unsupportedLangs: Set<string> }> {
-	let embedded = 0;
-	let skipped = 0;
-	const unsupportedLangs = new Set<string>();
-	const BATCH_SIZE = 50;
-
-	while (true) {
-		const unembedded = store.getUnembeddedMessages(chatId, BATCH_SIZE);
-
-		if (unembedded.length === 0) {
-			break;
-		}
-
-		for (const msg of unembedded) {
-			if (!msg.text || msg.text.trim().length < 3) {
-				skipped++;
-				continue;
-			}
-
-			try {
-				const langResult = await detectLanguage(msg.text);
-
-				if (!EMBEDDING_LANGUAGES.has(langResult.language)) {
-					unsupportedLangs.add(langResult.language);
-					skipped++;
-					continue;
-				}
-
-				const result: EmbedResult = await embedText(msg.text, langResult.language, "sentence");
-				const embedding = new Float32Array(result.vector);
-				store.insertEmbedding(chatId, msg.id, embedding);
-				embedded++;
-			} catch (err) {
-				logger.debug(`Embedding failed for message ${msg.id}: ${err}`);
-				skipped++;
-			}
-		}
-	}
-
-	return { embedded, skipped, unsupportedLangs };
-}
+// ── Embed Command ─────────────────────────────────────────────────────
 
 function registerEmbedCommand(history: Command): void {
 	history
@@ -456,7 +273,6 @@ function registerEmbedCommand(history: Command): void {
 					const { embedded, skipped, unsupportedLangs } = await embedMessages(
 						store,
 						contact.userId,
-						contact.displayName,
 					);
 
 					const total = store.getEmbeddedCount(contact.userId);
@@ -483,44 +299,6 @@ function registerEmbedCommand(history: Command): void {
 }
 
 // ── Search Command ────────────────────────────────────────────────────
-
-function formatSearchResult(result: SearchResult, contactName: string): string {
-	const msg = result.message;
-	const date = new Date(msg.date_unix * 1000);
-	const dateStr = date.toLocaleDateString("en-US", {
-		year: "numeric",
-		month: "short",
-		day: "numeric",
-		hour: "2-digit",
-		minute: "2-digit",
-	});
-	const direction = msg.is_outgoing ? pc.blue("You") : pc.cyan(contactName);
-	const text = msg.text || msg.media_desc || "(no text)";
-	const preview = text.length > 120 ? `${text.slice(0, 120)}...` : text;
-
-	let scoreLabel = "";
-
-	if (result.score !== undefined) {
-		scoreLabel = pc.dim(` [score: ${result.score.toFixed(4)}]`);
-	} else if (result.distance !== undefined) {
-		scoreLabel = pc.dim(` [dist: ${result.distance.toFixed(4)}]`);
-	} else if (result.rank !== undefined) {
-		scoreLabel = pc.dim(` [rank: ${result.rank.toFixed(2)}]`);
-	}
-
-	return `${pc.dim(dateStr)} ${direction}: ${preview}${scoreLabel}`;
-}
-
-function displayResults(results: SearchResult[], contactName: string): void {
-	if (results.length === 0) {
-		p.log.warn("No results found.");
-		return;
-	}
-
-	for (const result of results) {
-		p.log.info(formatSearchResult(result, contactName));
-	}
-}
 
 function registerSearchCommand(history: Command): void {
 	history
@@ -652,10 +430,8 @@ function registerExportCommand(history: Command): void {
 				output?: string;
 			},
 		) => {
-			const validFormats = ["json", "csv", "txt"];
-
-			if (!validFormats.includes(opts.format)) {
-				p.log.error(`Invalid format "${opts.format}". Use: ${validFormats.join(", ")}`);
+			if (!VALID_EXPORT_FORMATS.includes(opts.format as ExportFormat)) {
+				p.log.error(`Invalid format "${opts.format}". Use: ${VALID_EXPORT_FORMATS.join(", ")}`);
 				return;
 			}
 
@@ -690,46 +466,7 @@ function registerExportCommand(history: Command): void {
 				return;
 			}
 
-			let output: string;
-
-			switch (opts.format) {
-				case "json":
-					output = JSON.stringify(messages, null, 2);
-					break;
-
-				case "csv": {
-					const header = "id,date,sender,direction,text,media";
-					const rows = messages.map((m) => {
-						const direction = m.is_outgoing ? "sent" : "received";
-						const text = (m.text ?? "").replace(/"/g, '""').replace(/\n/g, "\\n");
-						const media = (m.media_desc ?? "").replace(/"/g, '""');
-						return `${m.id},"${m.date_iso}","${m.sender_id ?? ""}","${direction}","${text}","${media}"`;
-					});
-					output = [header, ...rows].join("\n");
-					break;
-				}
-
-				case "txt": {
-					const lines = messages.map((m) => {
-						const date = new Date(m.date_unix * 1000);
-						const dateStr = date.toLocaleString("en-US", {
-							year: "numeric",
-							month: "short",
-							day: "numeric",
-							hour: "2-digit",
-							minute: "2-digit",
-						});
-						const direction = m.is_outgoing ? "You" : contact.displayName;
-						const text = m.text || m.media_desc || "(no content)";
-						return `[${dateStr}] ${direction}: ${text}`;
-					});
-					output = lines.join("\n");
-					break;
-				}
-
-				default:
-					return;
-			}
+			const output = formatMessages(messages, opts.format as ExportFormat, contact.displayName);
 
 			if (opts.output) {
 				await Bun.write(opts.output, output);
